Split contract.rs into focused modules
contract.rs (703 lines) mixed submission, admin, governance, queries, and
helpers. Move it into src/contract/ with one module per concern —
submission.rs, admin.rs, governance.rs, queries.rs, utils.rs — keeping a
single #[contractimpl] block (the Soroban SDK generates one client and
spec per impl block) whose entry points delegate to the modules. The
public API surface and every signature are unchanged; all existing
crate::contract::* imports continue to work.
